service/wallet/requirements.txt
cryptography==40.0.2
Django==4.2
django-cors-headers==3.14.0
django-health-check==3.17.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.5
psycopg2-binary==2.9.6
